+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 4.0.3
+
 - [browser] feat: Better dedupe integration event description
 - [core] ref: Move Dedupe, FunctionString, InboundFilters and SdkInformation integrations to the core package
 - [core] feat: Provide correct platform and make a place to override event internals